Added watchMode option
