+++ lib/index.js
@@ -1,11 +1,10 @@
 const path = require('path');
-const chokidar = require('chokidar');
 const fs = require('fs-extra');
 const startShell = require('./utils/startShell');
 const setLogOutputToConsole = require('./utils/startShell').setLogOutputToConsole;
 const copyChaincode = require('./copyChaincode');
 const initChaincode = require('./initChaincode');
-const upgradeChaincode = require('./upgradeChaincode');
+const watchChaincode = require('./watchChaincode');
 const installNpmPackages = require('./installNpmPackages');
 const {SCRIPTS_PATH, LOCAL_CHAINCODE_PATH, DOCKER_SETUP_FINISHED_REGEX, DEFAULT_CHAINCODE_DESTINATION_PATH, DEFAULT_DOCKER_FILE} = require('./constants');
 
@@ -23,6 +22,7 @@ module.exports = function setupDevEnv({
         }
         return true;
     },
+    watchMode = true,
     dockerFile = DEFAULT_DOCKER_FILE,
     chaincodeDestination = DEFAULT_CHAINCODE_DESTINATION_PATH,
     copyGlobPattern = undefined
@@ -51,50 +51,10 @@ module.exports = function setupDevEnv({
         .then(installAllNpmPackages)
         .then(initAllChaincode)
         .then(() => {
-            // start watching
-
-            // Upgrade chaincode whenever source files change
-            chaincodeLocations.forEach((location, index) => {
-                // Chain the upgrade promises here ... else the chaincode becomes
-                // corrupt when quickly changing the same chaincode twice and it tries to upgrade
-                // 2 times asynchronously
-                let upgradePromise = Promise.resolve();
-                let previousUpgradeTimeout = null;
-
-                let version = 0;
-                chokidar.watch([
-                    path.join(location, '**/*.js'),
-                    `!${path.join(location, 'node_modules/**')}`
-                ], {ignoreInitial: true})
-                    .on('all', (event, path) => {
-                        // Debounce upgrades when updates are done to the same chaincode to quick
-                        if (previousUpgradeTimeout != null) {
-                            clearTimeout(previousUpgradeTimeout);
-                        }
-
-                        previousUpgradeTimeout = setTimeout(() => {
-                            previousUpgradeTimeout = null;
-
-                            upgradePromise = upgradePromise.then(() => {
-                                // Increase version number
-                                version += 1;
-    
-                                console.log(`Upgrading chaincode for ${location} to version ${version}`);
-                                // TODO: only copy changed file
-                                return copyChaincode(location, chaincodeDestination).then(() => {
-    
-                                    return upgradeChaincode(location, version)
-                                        .then(() => console.log(`Chaincode upgraded successful to version ${version}`))
-                                }).catch(err => {
-                                    console.log(`Chaincode upgrade to version ${version} failed`);
-                                    console.log(err);
-    
-                                    return Promise.resolve();
-                                });
-                            });
-                        }, 1000);
-                    });
-            });
+           if (watchMode) {
+               return watchChaincode(chaincodeLocations, chaincodeDestination);
+           }
+           return Promise.resolve();
         })
         .catch(() => {
             process.exit(1);
